+++ src/components/Input/Input.story.js
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, text, boolean, number, select } from '@storybook/addon-knobs';
 
-import Input from './index.js';
+import Input from './index';
 
 const stories = storiesOf('Input', module);
 stories